third_party/WebKit/LayoutTests/http/tests/security/promise-access-control-allow-expected.txt

CONSOLE MESSAGE: line 11: PASS: received unhandledrejection event
CONSOLE MESSAGE: line 13: PASS: received expected reason
Tests that promise rejection events are received for cross origin CORS scripts
